enforce prettier in eslint inspection
Also removes prettier call from package.json, handled by eslint-prettier-plugin.
This is the "best of both worlds" case, since it provides linting and inspection
for problematic code writing, but still doesn't provide accurate details on
which style rules are being violated.

This is better compared to solely using eslint-config-prettier, where absolutely
no feedback is provided to the developer regarding style issues. Such a thing
would be better if formatting works like autocrlf where the preferred version is
kept on the local machine, but that is unfortunately not the case.